Mv wfact into src, mv wfact tests to test
